+++ circuits/src/hash/sha256/mod.rs
@@ -104,26 +104,31 @@ mod tests {
 
     #[test]
     fn test_sha256_hash() {
+        fn test_wrapper(input_size: usize, k: u32, cost_model: bool) {
+            test_hash::<
+                Scalar,
+                AssignedByte<Scalar>,
+                [AssignedByte<Scalar>; 32],
+                Sha256Chip<Scalar>,
+                NativeGadget<Scalar, _, _>,
+            >(cost_model, "SHA256", input_size, k)
+        }
+
         const SHA256_BLOCK_SIZE: usize = 64;
-        let additional_sizes = [
-            SHA256_BLOCK_SIZE - 2,
-            SHA256_BLOCK_SIZE - 1,
-            SHA256_BLOCK_SIZE,
-            SHA256_BLOCK_SIZE + 1,
-            SHA256_BLOCK_SIZE + 2,
-            2 * SHA256_BLOCK_SIZE - 2,
-            2 * SHA256_BLOCK_SIZE - 1,
-            2 * SHA256_BLOCK_SIZE,
-            2 * SHA256_BLOCK_SIZE + 1,
-            2 * SHA256_BLOCK_SIZE + 2,
-        ];
-        test_hash::<
-            Scalar,
-            AssignedByte<Scalar>,
-            [AssignedByte<Scalar>; 32],
-            Sha256Chip<Scalar>,
-            NativeGadget<Scalar, _, _>,
-        >(true, "SHA256", &additional_sizes, 13);
+        const SHA256_EDGE_PADDING: usize = 55;
+        test_wrapper(2 * SHA256_BLOCK_SIZE, 14, true);
+
+        test_wrapper(SHA256_BLOCK_SIZE, 13, false);
+        test_wrapper(SHA256_BLOCK_SIZE - 1, 13, false);
+        test_wrapper(SHA256_BLOCK_SIZE - 2, 13, false);
+        test_wrapper(4 * SHA256_BLOCK_SIZE, 15, false);
+
+        test_wrapper(SHA256_EDGE_PADDING, 13, false);
+        test_wrapper(SHA256_EDGE_PADDING - 1, 13, false);
+
+        test_wrapper(0, 13, false);
+        test_wrapper(1, 13, false);
+        test_wrapper(2, 13, false);
     }
 
     #[test]